Patch to address potential integer overflow
Static analysis (coverity) detected that some of the math
in the routine getLatency was performed in 32 bit arithmetic,
even though the result was 64 bits. This could result in an
(theoretical) integer overflow. Cast to 64 bits to fix the
defect.

Note that a quick check (perhaps wrong) showed that this
method was never called, so this is probably an irrelevant
identified defect.

Fixes coverity 700584
(cherry picked from commit 499584ef7e20f97c738427fffc2657a56f1617ab)

Fixes #11595

+++ mythtv/libs/libmythfreesurround/freesurround.cpp
@@ -493,7 +493,7 @@ long long FreeSurround::getLatency()
     // returns in usec
     if (latency_frames == 0)
         return 0;
-    return decoder ? ((latency_frames + in_count)*1000000)/(2*srate) : 0;
+    return decoder ? ((long long)(latency_frames + in_count)*1000000)/(2*srate) : 0;
 }
 
 void FreeSurround::flush()
